some bugs

+++ main.h
@@ -13,7 +13,7 @@ int _printf(const char *format, ...);
 char *_strings(va_list, int);
 char *_char(va_list, int);
 char *_int_s(va_list, int);
-int i_sqr(int, int);
+long int i_sqr(int, int);
 int i_len(int);
 int _putch(int);
 int _putwrd(char *);
